included patch to update storage qty

+++ src/main/java/com/example/demo/model/Product.java
@@ -19,5 +19,5 @@ public class Product {
     private String id;
     private String name;
     private String type;
-    private int storage;
+    private double storage;
 }
